fix: eliminate all macOS build warnings
- Add printf format attribute to cli_color_fprintf so the compiler
  type-checks caller format strings and no longer flags the internal
  vfprintf as format-nonliteral
- Cast loop index explicitly before float multiply in physics_demo
- Silence Apple linker's duplicate-static-library warning, which is
  expected when targets link core both explicitly and transitively

+++ examples/physics_demo.c
@@ -9,7 +9,7 @@ int main(void) {
     printf("t=0: pos=(%.2f, %.2f, %.2f)\n", (double)p.pos.x, (double)p.pos.y, (double)p.pos.z);
     for (int i = 1; i <= 10; i++) {
         physics_step_euler(&p, 1, 0.1f, gravity);
-        printf("t=%.1f: pos=(%.2f, %.2f, %.2f)\n", (double)(i * 0.1f), (double)p.pos.x,
+        printf("t=%.1f: pos=(%.2f, %.2f, %.2f)\n", (double)((float)i * 0.1f), (double)p.pos.x,
                (double)p.pos.y, (double)p.pos.z);
     }
     return 0;
